extract temp module with tracking into own module
