+++ app/test/helpers/temp.ts
@@ -1,0 +1,7 @@
+/* eslint-disable no-sync */
+
+import * as temp from 'temp'
+const _temp = temp.track()
+
+export const mkdirSync = _temp.mkdirSync
+export const openSync = _temp.openSync
