+++ app/test/unit/validated-repository-path-test.ts
@@ -1,6 +1,7 @@
 /* eslint-disable no-sync */
 
-import { setupFixtureRepository, openSync } from '../helpers/repositories'
+import { setupFixtureRepository } from '../helpers/repositories'
+import { openSync } from '../helpers/temp'
 import { validatedRepositoryPath } from '../../src/lib/stores/helpers/validated-repository-path'
 
 describe('validatedRepositoryPath', () => {
